Update rasppi.md
line 144 to 145 added as an explianion of the problem not a way to correct it. only a few months using Linux compared to windows from dos 3.1 on different machines and windows 3.1 Windows for work goups 3.11. 

Not fully checked if it is that the useername is causing the problem. as just covered a rpi4 4gb and HDD 1TB with coke so not able to full check the problem.